backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - Using asyncpg for better Windows compatibility
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# AI/ML Libraries
transformers==4.46.0
torch>=2.6.0
tokenizers==0.21.0
datasets==3.2.0

# NLP and Explainability
shap==0.46.0
lime==0.2.0.1

# OCR
pytesseract==0.3.10
easyocr==1.7.0
Pillow==11.0.0
pdf2image==1.16.3

# Translation
sentencepiece==0.2.0
sacremoses==0.1.1

# Payments
razorpay==1.3.0

# Utilities
requests==2.32.3
numpy==2.2.1
pandas==2.2.3
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
